+++ tests/capstone2llvmir/arm64_tests.cpp
@@ -3714,6 +3714,38 @@ TEST_P(Capstone2LlvmIrTranslatorArm64Tests, ARM64_INS_LDXR)
 	EXPECT_NO_VALUE_CALLED();
 }
 
+TEST_P(Capstone2LlvmIrTranslatorArm64Tests, LdxrLoadIsAtomic)
+{
+	auto* f = translate(assemble("ldxr x0, [x1]"));
+	ASSERT_NE(nullptr, f);
+	bool found = false;
+	for (auto it = inst_begin(f), e = inst_end(f); it != e; ++it)
+	{
+		if (auto* l = dyn_cast<LoadInst>(&*it))
+		{
+			if (l->isAtomic())
+			{
+				found = true;
+				break;
+			}
+		}
+	}
+	EXPECT_TRUE(found);
+}
+
+TEST_P(Capstone2LlvmIrTranslatorArm64Tests, LdrLoadIsNotAtomic)
+{
+	auto* f = translate(assemble("ldr x0, [x1]"));
+	ASSERT_NE(nullptr, f);
+	for (auto it = inst_begin(f), e = inst_end(f); it != e; ++it)
+	{
+		if (auto* l = dyn_cast<LoadInst>(&*it))
+		{
+			EXPECT_FALSE(l->isAtomic());
+		}
+	}
+}
+
 //
 // ARM64_INS_LDXRB
 //